backend/requirements.txt
Flask==3.0.1
requests==2.31.0
transformers==4.38.2
torch==2.1.2
redis==5.0.1
python-dotenv==1.0.1
